# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_domain_loadbalance.CreateMemberRequest:
    properties:
      address:
        type: string
      pool_id:
        type: integer
      protocol_port:
        type: integer
      weight:
        type: integer
    type: object
  internal_domain_loadbalance.CreatePoolRequest:
    properties:
      description:
        type: string
      lb_method:
        type: string
      name:
        type: string
      protocol:
        type: string
    type: object
  internal_domain_loadbalance.CreateVIPRequest:
    properties:
      address:
        type: string
      connection_limit:
        type: integer
      description:
        type: string
      name:
        type: string
      pool_id:
        type: integer
      protocol:
        type: string
      protocol_port:
        type: integer
      session_persistence:
        type: boolean
    type: object
...
  internal_domain_loadbalance.Member:
    properties:
      address:
        type: string
      admin_state_up:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      pool_id:
        type: integer
      protocol_port:
        type: integer
      status:
        type: string
      weight:
        type: integer
    type: object
  internal_domain_loadbalance.Pool:
    properties:
      admin_state_up:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      lb_method:
        type: string
      name:
        type: string
      protocol:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_domain_loadbalance.UpdateMemberRequest:
    properties:
      address:
        type: string
      admin_state_up:
        type: boolean
      protocol_port:
        type: integer
      weight:
        type: integer
    type: object
  internal_domain_loadbalance.UpdatePoolRequest:
    properties:
      admin_state_up:
        type: boolean
      description:
        type: string
      lb_method:
        type: string
      name:
        type: string
      protocol:
        type: string
    type: object
  internal_domain_loadbalance.UpdateVIPRequest:
    properties:
      address:
        type: string
      admin_state_up:
        type: boolean
      connection_limit:
        type: integer
      description:
        type: string
      name:
        type: string
      pool_id:
        type: integer
      protocol:
        type: string
      protocol_port:
        type: integer
      session_persistence:
        type: boolean
    type: object
  internal_domain_loadbalance.VIP:
    properties:
      address:
        type: string
      admin_state_up:
        type: boolean
      connection_limit:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      pool_id:
        type: integer
      protocol:
        type: string
      protocol_port:
        type: integer
      session_persistence:
        type: boolean
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: WAF 防火墙管理系统控制面 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WAF Control API
  version: "1.0"
paths:
...
  /lb/pools:
    get:
      consumes:
      - application/json
      description: 查询所有负载均衡资源池
      produces:
      - application/json
      responses:
        "200":
          description: 资源池列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取资源池列表
      tags:
      - 负载均衡
    post:
      consumes:
      - application/json
      description: 创建新的负载均衡资源池
      parameters:
      - description: 资源池信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.CreatePoolRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Pool'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建资源池
      tags:
      - 负载均衡
  /lb/pools/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除负载均衡资源池
      parameters:
      - description: 资源池ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 资源池不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除资源池
      tags:
      - 负载均衡
    get:
      consumes:
      - application/json
      description: 根据ID获取负载均衡资源池详细信息
      parameters:
      - description: 资源池ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 资源池详情
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Pool'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 资源池不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取资源池详情
      tags:
      - 负载均衡
    put:
      consumes:
      - application/json
      description: 根据ID更新负载均衡资源池配置
      parameters:
      - description: 资源池ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.UpdatePoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Pool'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新资源池
      tags:
      - 负载均衡
...
  /lb/pools/{poolId}/members:
    get:
      consumes:
      - application/json
      description: 根据资源池ID获取该池下的所有成员
      parameters:
      - description: 资源池ID
        in: path
        name: poolId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成员列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取资源池成员列表
      tags:
      - 负载均衡
    post:
      consumes:
      - application/json
      description: 为指定资源池添加新的后端成员
      parameters:
      - description: 资源池ID
        in: path
        name: poolId
        required: true
        type: integer
      - description: 成员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.CreateMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Member'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建资源池成员
      tags:
      - 负载均衡
  /lb/pools/{poolId}/members/{memberId}:
    delete:
      consumes:
      - application/json
      description: 根据成员ID删除资源池成员
      parameters:
      - description: 资源池ID
        in: path
        name: poolId
        required: true
        type: integer
      - description: 成员ID
        in: path
        name: memberId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 成员不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除资源池成员
      tags:
      - 负载均衡
    get:
      consumes:
      - application/json
      description: 根据成员ID获取成员详细信息
      parameters:
      - description: 资源池ID
        in: path
        name: poolId
        required: true
        type: integer
      - description: 成员ID
        in: path
        name: memberId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成员详情
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Member'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 成员不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取资源池成员详情
      tags:
      - 负载均衡
    put:
      consumes:
      - application/json
      description: 根据成员ID更新成员配置
      parameters:
      - description: 资源池ID
        in: path
        name: poolId
        required: true
        type: integer
      - description: 成员ID
        in: path
        name: memberId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.Member'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新资源池成员
      tags:
      - 负载均衡
  /lb/vips:
    get:
      consumes:
      - application/json
      description: 查询所有虚拟IP地址
      produces:
      - application/json
      responses:
        "200":
          description: VIP列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取VIP列表
      tags:
      - 负载均衡
    post:
      consumes:
      - application/json
      description: 创建新的虚拟IP地址
      parameters:
      - description: VIP信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.CreateVIPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.VIP'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建VIP
      tags:
      - 负载均衡
  /lb/vips/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除虚拟IP地址
      parameters:
      - description: VIP ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: VIP不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除VIP
      tags:
      - 负载均衡
    get:
      consumes:
      - application/json
      description: 根据ID获取虚拟IP地址详细信息
      parameters:
      - description: VIP ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: VIP详情
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.VIP'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: VIP不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取VIP详情
      tags:
      - 负载均衡
    put:
      consumes:
      - application/json
      description: 根据ID更新虚拟IP地址配置
      parameters:
      - description: VIP ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_loadbalance.UpdateVIPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_domain_loadbalance.VIP'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新VIP
      tags:
      - 负载均衡
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"